+++ packages/data-loader/src/parsers/parseCsv/__tests__/groupByIndex.test.ts
@@ -1,4 +1,4 @@
-import { RECORD_INDEX } from "../../../printers/csvPrinter/constants";
+import { RECORD_INDEX } from "../../../printers/printAsCsv/constants";
 import { groupByIndex } from "../groupByIndex";
 
 describe("groupByIndex", () => {
