+++ packages/data-loader/src/parsers/parseCsv/extractSubTableFieldsValue.ts
@@ -1,4 +1,4 @@
-import { FieldsJson } from "../../printers/csvPrinter";
+import { FieldsJson } from "../../printers/printAsCsv";
 import { KintoneFormFieldProperty } from "@kintone/rest-api-client";
 import { formatToRecordValue } from "./formatToRecordValue";
 import { CsvRecords } from "./index";
