+++ packages/data-loader/src/parsers/parseCsv/index.ts
@@ -1,8 +1,8 @@
 import csvParse from "csv-parse/lib/sync";
 import { KintoneFormFieldProperty } from "@kintone/rest-api-client";
-import { PRIMARY_MARK } from "../../printers/csvPrinter/constants";
-import { FieldsJson } from "../../printers/csvPrinter";
-import { hasSubTable } from "../../printers/csvPrinter/hasSubTable";
+import { PRIMARY_MARK } from "../../printers/printAsCsv/constants";
+import { FieldsJson } from "../../printers/printAsCsv";
+import { hasSubTable } from "../../printers/printAsCsv/hasSubTable";
 import { extractSubTableFieldsValue } from "./extractSubTableFieldsValue";
 import { groupByIndex } from "./groupByIndex";
 import { isImportSupportedFieldType } from "./isImportSupportedFieldType";
