+++ packages/data-loader/src/parsers/parseJson.ts
@@ -1,4 +1,4 @@
-import { ParsedRecord } from "./csvParser";
+import { ParsedRecord } from "./parseCsv";
 
 export const parseJson = (jsonString: string) => {
   return JSON.parse(jsonString) as ParsedRecord[];
