+++ packages/data-loader/src/printers/printAsCsv/index.ts
@@ -9,7 +9,7 @@ export type FieldsJson = {
   properties: { [k: string]: KintoneFormFieldProperty.OneOf };
 };
 
-export const csvPrinter = (
+export const printAsCsv = (
   records: KintoneRecord[],
   fieldsJson: FieldsJson
 ) => {
