+++ packages/data-loader/src/printers/printAsJson.ts
@@ -1,6 +1,6 @@
 import { KintoneRecordField } from "@kintone/rest-api-client";
 
-export const jsonPrinter = (
+export const printAsJson = (
   records: Array<{ [k: string]: KintoneRecordField.OneOf }>
 ) => {
   console.log(JSON.stringify(records, null, 2));
